test: move sync config tests to integration file
